+++ src/main/java/sitemesh/SiteMeshFilter.java
@@ -19,11 +19,9 @@ public class SiteMeshFilter extends ConfigurableSiteMeshFilter {
 		.addExcludedPath("/users/updateForm")
 		.addExcludedPath("/users/pwForm")
 		.addExcludedPath("/board/deleteForm")
-		;
+		.addExcludedPath("/users/notificationForm");
 
-		builder.addDecoratorPath("/mainLMS/*", "/layouts/MainLayout.jsp")
-		.addExcludedPath("/mainLMS/notificationForm");
-		
+		builder.addDecoratorPath("/mainLMS/*", "/layouts/MainLayout.jsp");
 		builder.addDecoratorPath("/deptLMS/*", "/layouts/MainLayout.jsp");
 		builder.addDecoratorPath("/classLMS/*", "/layouts/MainLayout.jsp");
 		// 동원 수정
